Created models package and moved HelloObject to models
