+++ src/main/java/edu/neu/cs4500/models/HelloObject.java
@@ -1,4 +1,4 @@
-package edu.neu.cs4500.controllers.hello;
+package edu.neu.cs4500.models;
 
 public class HelloObject {
